Fix wrong slot being selected if SRAM is cleared
This can happen when a user renames a new rom to a previous one (e.g. for MSU packs) and there is existing sram data that is cleared out when the checksum doesn't match. the save slot selected marker and its complement were not being cleared out in the clearSRAM routine meaning the default selected slot could be 2 or 3.